+++ changelog.md
@@ -1,4 +1,10 @@
-# Working with templates
+### Working with form
+- Validating forms
+- Receiving data from the form in the controller
+- Added @csrf directive into the form
+- Added route to respond to post method.
+- Added form for contact
+### Working with templates
 - Separate the navigation from the layout
 - Changed composer autoloader, must recompile the autoload `composer dumpautoload`
 - Added helper class function